Moved game mechanics to comp organizaiton: shooting system, movement controls, collision detection, and target handling;, score display and weapon types, LaserBeam component and time utility.
