+++ app/pages/asteroid/_comp/Weapons/constants.js
@@ -1,0 +1,5 @@
+export const WEAPON_TYPES = [
+  { key: 'spread', name: 'Spread Shot', maxAmmo: 30, cooldown: 0.3 },
+  { key: 'laser', name: 'Laser Beam', maxAmmo: 50, cooldown: 0 },
+  { key: 'explosive', name: 'Explosive Shot', maxAmmo: 5, cooldown: 2 },
+];
