+++ app/utils/time.js
@@ -1,0 +1,1 @@
+export const now = () => performance.now() / 1000;
